Fixed issue with BG img & links
The links in about section of home page stopped working with BG image - but it wasn't the image. It was a layer above the image to darken it. Removed the layer and adjusted the image in PS. Was thinking a % adjustment ability in code was most flexible approach.... oh well, I'm sure there's a way I don't know about.